pyproject.toml
[tool.poetry]
name = "cool_project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.7.13"
torch = "^1.12"
torchvision = "^0.13.0"
tqdm = "^4.64.0"
numpy = "1.21"
rich = "^12.5.1"
loguru = "^0.6.0"
pytorch-lightning = "^1.7.1"
pydantic = "^1.9.2"
python-dotenv = "^0.20.0"
torchmetrics = "^0.9.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
black = "^22.3.0"
mypy = "^0.961"
pycln = "^2.0.1"
isort = "^5.10.1"
pre-commit = "^2.19.0"
pre-commit-hooks = "^4.3.0"
jupyter = "^1.0.0"
ipykernel = "^6.15.1"

[tool.mypy]
files = ["^cool_project/"]
pretty = true
strict = true
# show_error_codes = true
enable_error_code = ["ignore-without-code", "redundant-expr", "truthy-bool"]
# ignore_missing_imports = true
exclude = "(/tests/)|(/docs/)|(/examples/)"

[tool.pytest]
# log_cli = true
addopts = [
    # show summary of all tests that did not pass
    "-ra",
    # idk
    "-s",
    # Make tracebacks shorter
    "--tb=native",
    # enable all warnings
    "-Wd",
]
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.isort]
profile = "black"
line_length = 79

[tool.black]
line-length = 79 # override the default of 88

[build-system]
requires = ["poetry>=1.1.6"]
build-backend = "poetry.masonry.api"
